% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_metadata_shiny}
\alias{add_metadata_shiny}
\title{Select files for flow object - Shiny}
\usage{
add_metadata_shiny(flow_object)
}
\arguments{
\item{flow_object}{working directory}
}
\value{
A file list
}
\description{
Select files
}
