% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_sample_metadata}
\alias{add_sample_metadata}
\title{Add metadata annotations}
\usage{
add_sample_metadata(
  flow_object,
  metadata = NULL,
  key_column = NULL,
  interactive = FALSE
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{metadata}{A data.frame containing sample annotations.}

\item{key_column}{Name of the column matching sample names.}

\item{interactive}{Logical argument to determine whether or not to activate the interactive mode, which opens up a graphical interface to select a metadata file. Defaults to FALSE.}
}
\value{
A Flow Object
}
\description{
Add sample annotations to the Flow Object. This is where you add clinical variables (i.e. antibody levels, Timepoints, Treatment, etc) or technical variables (like Batch).
Variables can be discrete (i.e. categorical) or continuous. If using categorical, it is recommended to assign factor levels if you desire a specific order for labels to be displayed in plots, or how statistics will be performed.
By default, the order will be performed in alphabetical order.
}
