% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_staining_controls}
\alias{add_staining_controls}
\title{Add Staining controls}
\usage{
add_staining_controls(
  flow_object,
  interactive = T,
  control_list = NULL,
  downsample = 2000
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{interactive}{Activates the interactive mode: opens a graphical interface to select samples instead of providing a control_list. Defaults to TRUE.}

\item{control_list}{A named list where names correspond to markers found in the Flow Object, and values are paths to FCS files. At least one value must also be for an Unstained Universal Control. see Details.}

\item{downsample}{Number of events to select from each FCS file.}
}
\value{
A FlowSet
}
\description{
Select control FCS files as references for plotting.
}
