% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_staining_controls_shiny}
\alias{add_staining_controls_shiny}
\title{Shiny interface to add reference controls}
\usage{
add_staining_controls_shiny(markers)
}
\arguments{
\item{markers}{A Flow Object}
}
\value{
A named list of reference control files.
}
\description{
Boot the GUI file selector
}
