% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{alpha_hull_find}
\alias{alpha_hull_find}
\title{Select an alpha value for an alpha convex hull for a given trajectory}
\usage{
alpha_hull_find(plot)
}
\arguments{
\item{plot}{A Pseudotime UMAP plot. generated by the `impute_pseudotime() function`}
}
\value{
An alpha value
}
\description{
Select an alpha value for an alpha convex hull for a given trajectory
}
