% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{Annotate louvain clusters}
\usage{
annotate_clusters(flow_object, cluster_annotations = NULL, interactive = FALSE)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{cluster_annotations}{A data.frame containing a column with louvain labels, and additional columns with user-defined labels. Ignored if using the 'interactive' mode.}

\item{interactive}{Logical argument to determine whether or not to activate the interactive mode, which opens up a graphical interface to manually add labels. Defaults to FALSE.}
}
\value{
A flow_object
}
\description{
Add annotations to louvain clusters. Multiple concurrent annotations can be used in parallel. For instance, you could have more basic cell types `(CD4 T cells, CD8 T cells, etc)`, or more granular.
It is possible to have multiple clusters with the same annotations if you wish to group them. Those annotations will then be useable in plots and statistics instead of the default louvain annotations.
}
