% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotate_clusters_shiny}
\alias{annotate_clusters_shiny}
\title{Louvain annotation shiny interface}
\usage{
annotate_clusters_shiny(flow_object)
}
\arguments{
\item{flow_object}{A Flow Object}
}
\value{
A data.frame
}
\description{
Boots the annotate_cluster shiny interface. It is meant to be called by annotate_clusters function, not directly.
}
