% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd4_flow-data.R
\docType{data}
\name{cd4_flow}
\alias{cd4_flow}
\title{CD4 T cell flow object}
\format{
An object of class \code{flow_object} of length 6.
}
\usage{
data("cd4_flow")
}
\description{
Data from a pregated polyfunctional CD4 T cell staining `(CD3+CD4+)`.
}
\keyword{datasets}
