% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{Cluster_flow}
\alias{Cluster_flow}
\title{Phenograph Clustering}
\usage{
Cluster_flow(flow_object, k = 60, scale = TRUE)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{k}{number of neighbors for Phenograph. Defaults to 60.}

\item{scale}{Logical argument to determine whether or not to scale the data prior to umap. Defaults to TRUE}
}
\value{
A Flow Object
}
\description{
Runs Phenograph clustering on the flow data to identify louvain clusters from a flow object with embedded UMAP coordinates.
Uses the same transformation used for `UMAP_flow()`.
}
