% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cluster_jitterplot}
\alias{cluster_jitterplot}
\title{Plot Cluster jitter plot}
\usage{
cluster_jitterplot(
  flow_object,
  annotation = "louvain",
  value = "relative",
  louvain_select = NULL,
  group = NULL,
  stats = NULL
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{annotation}{Cluster annotation to display. Defaults to `louvain`.}

\item{value}{Which measure of abundance to display on the plot: can acdept either `relative` or `absolute` frequencies. Defaults to `relative`.}

\item{louvain_select}{Filter on a subset of louvain clusters if given a vector of clusters. `NULL`by default.}

\item{group}{Pass a metadata variable to compare frequencies across. Defaults to `NULL`.}

\item{stats}{Selection of a statistical test to compare groups (if using the `groups` parameter). Accepts `wilcox`, `t.test` or `none`. Defaults to `wilcox`.}
}
\value{
A ggplot object
}
\description{
Graphs a jitter plot of the frequency of cells in each sample for each cluster/cluster annotation.
This can then be set to be grouped by a metadata variable
}
