% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{clusters_compare_groups}
\alias{clusters_compare_groups}
\title{Discrete analysis of cluster frequencies}
\usage{
clusters_compare_groups(
  flow_object,
  outcome,
  method = "non-parametric",
  group.by = "louvain"
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{outcome}{Discrete or categorical metadata variable or outcome. Can be of 'character' or 'factor' class.}

\item{method}{Correlation test to use: either 'wilcox' or 't.test'. Defaults to `wilcox`. Columns with more than 2 levels of categories will use the 'pairwise' versions fo the statistical tests.}

\item{group.by}{Which cluster annotation to compute with: defaults to 'louvain'}
}
\value{
A data.frame
}
\description{
Statistical comparison of cluster frequencies versus metadata discrete (i.e. categorical) variables. Metadata needs to be added using the `add_sample_metadata` function.
}
