% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ClusterSubsampling}
\alias{ClusterSubsampling}
\title{Subsample on a per cluster basis}
\usage{
ClusterSubsampling(flow_object, n = 10000)
}
\arguments{
\item{flow_object}{A flow_object}

\item{n}{number of cells to subsample per clusters. Defaults to 10000}
}
\value{
A flow_object
}
\description{
Subset a FlowObject on metadata criteria, cluster annotations, dimensions or marker fluorescence.
}
