% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{correlate_clusters}
\alias{correlate_clusters}
\title{Regression analysis of cluster frequencies}
\usage{
correlate_clusters(
  flow_object,
  outcome,
  method = "spearman",
  group.by = "louvain"
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{outcome}{Numerical metadata variable or outcome}

\item{method}{Correlation test to use: either 'spearman' or 'pearson'. Defaults to `spearman`.}

\item{group.by}{Which cluster annotation to compute with: defaults to 'louvain'}
}
\value{
A data.frame
}
\description{
Statistical comparison of cluster frequencies versus metadata continuous variables. Metadata needs to be added using the `add_sample_metadata` function.
}
