% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{CreateFlowObject}
\alias{CreateFlowObject}
\title{Create a FlowObject from flow data}
\usage{
CreateFlowObject(
  flowSet = NULL,
  files = NULL,
  name = NULL,
  subsample = 5000,
  min_events = 500,
  max_total_events = 2e+05,
  unequal = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{flowSet}{A FlowSet object. Defaults to NULL. Either a flowSet or files must be supplied using the `files` parameter.}

\item{files}{A list of flow cytometry sample files. Can be either CSV or FCS files exported by FlowJo. Either a flowSet or files must be supplied using the `flowSet` parameter.}

\item{name}{identifier for the dataset}

\item{subsample}{Number of events to subsample per sample. Defaults to 5000.}

\item{min_events}{Minimum number of events per file. Any sample found with events under this value will be ignored. Defaults to 500.}

\item{max_total_events}{Total number of events across samples cannot exceed this value: subsampling number will be adjusted accordingly. Defaults to  2e+05}

\item{unequal}{Logical argument to determine whether or not to allow to take unequal number of events if some samples are limiting. When set to True. Defaults to FALSE.}

\item{interactive}{Logical argument to determine whether or not to use the interactive mode to select files in a graphical user interface . Defaults to FALSE.}
}
\value{
A Flow Object
}
\description{
Randomly selects n events from each file in either a flowSet or a list of FCS files.
}
\examples{
library(flowCore)
data(GvHD)
flow_obj <- CreateFlowObject(flowSet = GvHD, name = "GvHD")

}
