% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csv_to_flowSet}
\alias{csv_to_flowSet}
\title{Convert a list of CSV files exported from FlowJo to a flowSet}
\usage{
csv_to_flowSet(files)
}
\arguments{
\item{files}{a list of csv files}
}
\value{
A FlowSet
}
\description{
Randomly selects n events from each file in a subset, in order to get fewer events than the total number across all files.
}
