% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Downsampling_FlowSet}
\alias{Downsampling_FlowSet}
\title{Downsample a FlowSet to keep n events in each file}
\usage{
Downsampling_FlowSet(flowSet, n)
}
\arguments{
\item{flowSet}{A FlowSet}

\item{n}{A number of events to take per file. Values between 100 and 500000 are accepted.}
}
\value{
A FlowSet
}
\description{
Randomly selects n events from each file in a subset, in order to get fewer events than the total number across all files.
}
\examples{
library(flowCore)
data("GvHD")
Downsampling_FlowSet(flowSet = GvHD, n = 1000)

}
