% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{exclude_parameters}
\alias{exclude_parameters}
\title{Exclude markers}
\usage{
exclude_parameters(flow_object, exclude_markers)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{exclude_markers}{A vector of markernames to exclude.}
}
\value{
A Flow Object
}
\description{
Excludes markers in order to avoid using them for dimension reduction, clustering and other techniques.
To be used if cells were pregated on, or if some markers are unreliable.
}
