% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting.R
\name{export_to_fcs}
\alias{export_to_fcs}
\title{Export a Flow Object to a FCS file}
\usage{
export_to_fcs(flow_object, filename)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{filename}{A filename for the output FCS file. Must contain the `.fcs` extension.}
}
\value{
A FCS file
}
\description{
Generates a single FCS file, that can be imported into FlowJo, with embedded cluster labels, UMAP coordinates, and sample metadata.
Columns with character values, like cluster annotations or sample annotations, will be encoded as numbers, and a key to those values will be generated as a seaparate CSV file.
}
