% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_cofactors}
\alias{find_cofactors}
\title{Find optimal cofactors for hyperbolic sine transform}
\usage{
find_cofactors(flow_object)
}
\arguments{
\item{flow_object}{A Flow Object}
}
\value{
A flow_object
}
\description{
Incrementally increase cofactor values for an hyperbolic sine transformation for each marker until the fluorescence values are not longer split into a bimodal dsitribution centered around zero.
This is key to avoid having clustering and dimension reduction driven by very low expression markers.
}
