% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{FindTrajectoryMarkers}
\alias{FindTrajectoryMarkers}
\title{Plot a pseudotime marker heatmap}
\usage{
FindTrajectoryMarkers(cds, trajectory)
}
\arguments{
\item{cds}{A Flow Object}

\item{trajectory}{Name of the trajectory to plot.}
}
\value{
A ComplexHeatmap
}
\description{
Display the relative expression for each  marker relative to pseudoplot.
}
