% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{flow_object_to_cds}
\alias{flow_object_to_cds}
\title{Export a Flow Object to a CDS (cell_data_set).}
\usage{
flow_object_to_cds(flow_object, scale = TRUE)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{scale}{Logical argument to determine whether or not to use scaled data. Default is TRUE .}
}
\value{
A spreadsheet file
}
\description{
Generates a CDS (Monocle3) Object from a Flow Object.Internal function.
}
