% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowobj-data.R
\docType{data}
\name{flowobj}
\alias{flowobj}
\title{CD4 T cell flow object}
\format{
An object of class \code{flow_object} of length 6.
}
\usage{
data(flowobj)
}
\description{
Data from a pregated polyfunctional CD4 T cell staining `(CD3+CD4+)`.
}
\keyword{datasets}
