% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Get_ClusterAnnotations}
\alias{Get_ClusterAnnotations}
\title{Get Marker Names}
\usage{
Get_ClusterAnnotations(flow_object)
}
\arguments{
\item{flow_object}{a Flow Object}
}
\value{
A marker name vector
}
\description{
Retrieves marker Names from a Flow Object
}
