% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_flowSet_matrix}
\alias{get_flowSet_matrix}
\title{Get fluorescence data}
\usage{
get_flowSet_matrix(flow_object, add_sample_id = FALSE, annotations = FALSE)
}
\arguments{
\item{flow_object}{A flow_object}

\item{add_sample_id}{Logical argument to determine whether or not to add a SampleID column for each cell}

\item{annotations}{Logical argument to determine whether or not to add additional annotations to the data, including clusters, cluster annotations and sample metadata, and dimensions reductions.}
}
\value{
A data.frame
}
\description{
Extract the expression matrix from a flow_object, which can be complemented with metadata.
}
\examples{
library(flowCore)
data("GvHD")
flow_obj <- CreateFlowObject(flowSet = GvHD, name = "GvHD")
exprs <- get_flowSet_matrix(flow_obj, add_sample_id = TRUE)
}
