% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Get_MarkerNames}
\alias{Get_MarkerNames}
\title{Get Marker Names}
\usage{
Get_MarkerNames(flow_object, select = "all", show_channel_name = TRUE)
}
\arguments{
\item{flow_object}{a Flow Object}

\item{select}{Put a filter on which markers to display. Accepted values are 'all', 'included' or 'excluded'.}

\item{show_channel_name}{Logical argument to determine whether or not to also display associated channel names. Defaults to TRUE.}
}
\value{
A marker name vector
}
\description{
Retrieves marker Names from a Flow Object
}
