% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Get_MetaData}
\alias{Get_MetaData}
\title{Get Sample Metadata}
\usage{
Get_MetaData(flow_object)

Get_MetaData(flow_object)
}
\arguments{
\item{flow_object}{a Flow Object}
}
\value{
A data.frame

A data.frame
}
\description{
Retrieves sample metadata from a Flow Object

Retrieves cluster annotations from a Flow Object
}
