% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{import_trajectory}
\alias{import_trajectory}
\title{Import trajectories and pseudotime from CDS to a Flow Object}
\usage{
import_trajectory(flow_object, cds, trajectory_name)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{cds}{a 'cell_data_set' object to import.}

\item{trajectory_name}{Name of the trajectory to add to the flow object.}
}
\value{
A Flow Object with pseudotime values for given trajectory
}
\description{
Import trajectories and pseudotime from CDS to a Flow Object
}
