% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{impute_pseudotime}
\alias{impute_pseudotime}
\title{Impute pseudotime values for cells within a trajectory.}
\usage{
impute_pseudotime(flow_object, trajectory, k = 60, scale = TRUE)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{trajectory}{Name of the trajectory to add to the flow object.}

\item{k}{Number of neighbors to impute pseudotime values from. Defaults to 60.}

\item{scale}{Logical argument to determine whether or not to scale data to determine nearest neighbours. Recommended to use the same setting used in `UMAP_flow()`. Defaults to TRUE.}
}
\value{
A Flow Object with imputed pseudotime values for given trajectory
}
\description{
Cluster based downsampling facilitates computation of trajectories, with an obvious caveat: it only yield pseudotime values
on events that were kept by downsampling. If you wish to then compare pseudotime values across metadata groups, frequencies are no
longer reflecting true frequencies. This function first lets you isolate the location of a given trajectory using an alpha convex hull
algorithm by selecting an alpha value that best captures all events in the trajectory, then imputes pseudotime on missing values
within that defined area by getting the mean of k neighbors.
}
