% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_heatmap}
\alias{plot_cluster_heatmap}
\title{Plot Cluster Heatmap}
\usage{
plot_cluster_heatmap(
  flow_object,
  annotation = NULL,
  show_control = FALSE,
  min_mfi = -1000,
  clustering_method = "ward.D2",
  clustering_distance = "euclidean",
  louvain_select = NULL,
  add_mfi = FALSE,
  annotation_df = NULL
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{annotation}{A cluster annotation to display.}

\item{show_control}{Logical argument to determine whether or not show reference control MFI to the heatmap. Defaults to FALSE.}

\item{min_mfi}{Minimum MFI value to be displayed on the heatmap. Any value below this will be set to this value. Useful if you have a few extremely negative events. Defaults to -1000.}

\item{clustering_method}{Clustering method for hierarchical clustering. Viable options can be found with ?hclust}

\item{clustering_distance}{Clustering distance for hierarchical clustering. Viable options can be found with ?dist}

\item{louvain_select}{Select louvain clusters to represent on the heatmap}

\item{add_mfi}{Logical argument to determine whether or not to add the MFI representation of each cluster on the absolute scale when set to TRUE.}

\item{annotation_df}{Provide a custom cluster annotation data.frame.}
}
\value{
A ggplot object
}
\description{
Graphs a histogram, similar to a 1D FlowJo plot, of the expression of a selected marker for each cluster in a Flow Object.
}
