% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_histograms}
\alias{plot_cluster_histograms}
\title{Plot Cluster histogram}
\usage{
plot_cluster_histograms(
  flow_object,
  marker,
  clusters = NULL,
  limits = c(-100, 4e+05),
  alpha = 0.5,
  show_control = FALSE
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{marker}{Select which marker to make histograms for.}

\item{clusters}{Filter on a subset of louvain clusters if given a vector of clusters. NULL by default.}

\item{limits}{Sets the lower and upper limits of the x axis. Defaults to `c(-100, 400000)`}

\item{alpha}{Control point transparency, 1 is opaque, 0 is invisible. Defaults to 0.5}

\item{show_control}{Determine if the histogram for associated control is added to the plot. Requires prior addition of controls with `add_staining_controls()`. Defaults to FALSE}
}
\value{
A ggplot object
}
\description{
Graphs a histogram, similar to a 1D FlowJo, plot of the expression of a selected marker for each cluster in a Flow Object.
}
