% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimplots.R
\name{plot_density_umap}
\alias{plot_density_umap}
\title{UMAP Density Plot}
\usage{
plot_density_umap(
  flow_object,
  dot_size = 0.05,
  alpha = 0.5,
  split.by = NULL,
  ncols = NULL,
  densityn = 256,
  show_trajectories = FALSE
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{dot_size}{Control size of points. higher numbers for larger dots. Defaults to 0.5.}

\item{alpha}{Control point transparency, 1 is opaque, 0 is invisible. Defaults to 0.5}

\item{split.by}{Split the plot by sample/cluster annotations, similar to group.by. Can split by up to 3 variables.}

\item{ncols}{Controls the number of columns of plot if splitting by one variable}

\item{densityn}{Controls the density controller setting}

\item{show_trajectories}{Logical argument to determine whether or not to display computed trajectories on the plot. Defaults to FALSE.}
}
\value{
A ggplot object
}
\description{
Graphs the output of a UMAP dimension reduction on a 2D scatter plot where each point is an event,
positioned based on cell embeddings determined by the `RunUMAP()` function. Cells are colored by the local density.
Can be split into metadata annotation to compare the relative distribution of cells across biological conditions qualitatively.
}
\seealso{
\code{\link{plot_umap}} \code{\link{plot_umap_heat}} \code{\link{plot_umap_pseudotime}}
}
