% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{plot_pseudotime_boxplot}
\alias{plot_pseudotime_boxplot}
\title{Plot a pseudotime boxplot per group}
\usage{
plot_pseudotime_boxplot(flow_object, trajectory, group.by)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{trajectory}{Name of the trajectory to plot.}

\item{group.by}{Sample or cluster annotation to separate cells by.}
}
\value{
A ggplot2 object
}
\description{
Display a boxplot for pseudotime on the Y-axis, cluster or sample annotations on the X-Axis.
}
