% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{plot_pseudotime_markers}
\alias{plot_pseudotime_markers}
\title{Plot a pseudotime boxplot per group}
\usage{
plot_pseudotime_markers(flow_object, trajectory, method = "ward.D2")
}
\arguments{
\item{flow_object}{A Flow Object}

\item{trajectory}{Name of the trajectory to plot.}

\item{method}{Clustering method to arrange markers. Defaults to "ward.D2".  Accepted values can be found with ?hclust.}
}
\value{
A ComplexHeatmap
}
\description{
Display a boxplot for pseudotime on the Y-axis, cluster or sample annotations on the X-Axis.
}
