% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimplots.R
\name{plot_umap}
\alias{plot_umap}
\title{UMAP Dimensional reduction plot}
\usage{
plot_umap(
  flow_object,
  dot_size = 0.5,
  alpha = 0.5,
  group.by = NULL,
  split.by = NULL,
  ncols = NULL,
  label_filter = NULL,
  label = TRUE,
  show_trajectories = FALSE
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{dot_size}{Control size of points. higher numbers for larger dots. Defaults to 0.5.}

\item{alpha}{Control point transparency, 1 is opaque, 0 is invisible. Defaults to 0.5}

\item{group.by}{Color points grouping variable. Can include 1. louvain labels; 2. cluster annotations if previously provided; 3. metadata}

\item{split.by}{Split the plot by sample/cluster annotations, similar to group.by. Can split by up to 3 variables.}

\item{ncols}{Controls the number of columns of plot if splitting by one variable}

\item{label_filter}{Subset louvain label filter. Useful if you only wish to show a subset of labels.}

\item{label}{Logical argument to determine whether or not to activate labelling on the plot. FALSE turns on legend. Defaults to TRUE.}

\item{show_trajectories}{Logical argument to determine whether or not to display computed trajectories on the plot. Defaults to FALSE.}
}
\value{
A ggplot object
}
\description{
Graphs the output of a UMAP dimension reduction on a 2D scatter plot where each point is an event,
positioned based on cell embeddings determined by the `RunUMAP()` function. By default, cells are not colored,
but can be mapped to louvain clusters, sample or cluster annotations using the group.by parameter.
}
\seealso{
\code{\link{plot_density_umap}} \code{\link{plot_umap_heat}} \code{\link{plot_umap_pseudotime}}
}
