% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimplots.R
\name{plot_umap_heat}
\alias{plot_umap_heat}
\title{UMAP Marker Fluorescence Plot}
\usage{
plot_umap_heat(
  flow_object,
  colorGradient = c("blue4", "blue", "cyan", "green", "yellow", "orange", "red"),
  markers,
  dot_size = 0.5,
  alpha = 1,
  ncols = NULL,
  nrows = NULL,
  legend_height = 10,
  show_trajectories = FALSE,
  quantile_saturation = 1000
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{colorGradient}{A vector of colors to use as a gradient for expression, from low to high. Defaults to a rainbow palette.}

\item{markers}{Marker(s) to use to map expression values. A minimum of one marker is required.}

\item{dot_size}{Control size of points. higher numbers for larger dots. Defaults to 0.5.}

\item{alpha}{Control point transparency, 1 is opaque, 0 is invisible. Defaults to 1}

\item{ncols}{Number of columns to facet multiple plots by. Applicable if using more than 1 marker.}

\item{nrows}{Number of rows to facet multiple plots by. Applicable if using more than 1 marker.}

\item{legend_height}{Sets the height of the legend. Useful if using multiple markers. Defaults to 10.}

\item{show_trajectories}{Logical argument to determine whether or not to display computed trajectories on the plot. Defaults to FALSE.}

\item{quantile_saturation}{Uses quantile based saturation to limit the effect of rare very bright events can have the the color scale as a whole. Tweak this parameter depending on the rarity of those bright events.}
}
\value{
A ggplot object
}
\description{
Graphs the output of a UMAP dimension reduction on a 2D scatter plot where each point is an event,
positioned based on cell embeddings determined by the `RunUMAP()` function. By default, cells are not colored,
but can be mapped to louvain clusters, sample or cluster annotations using the group.by parameter.
}
\seealso{
\code{\link{plot_umap}} \code{\link{plot_density_umap}} \code{\link{plot_umap_pseudotime}}
}
