% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimplots.R
\name{plot_umap_pseudotime}
\alias{plot_umap_pseudotime}
\title{UMAP Pseudotime Plot}
\usage{
plot_umap_pseudotime(
  flow_object,
  dot_size = 0.5,
  alpha = 1,
  trajectory,
  viridis_palette = "viridis"
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{dot_size}{Control size of points. higher numbers for larger dots. Defaults to 0.5.}

\item{alpha}{Control point transparency, 1 is opaque, 0 is invisible. Defaults to 0.5}

\item{trajectory}{Name of the trajectory from which to display pseudotime.}

\item{viridis_palette}{Name of the viridis palette to use to color pseudotime values.}
}
\value{
A ggplot object
}
\description{
Graphs the output of a UMAP dimension reduction on a 2D scatter plot where each point is an event,
positioned based on cell embeddings determined by the `RunUMAP()` function. Cells are colored by the pseudotime value along the chosen trajectory.
Requires a trajectory analysis to be first performed on the flow object.
}
\seealso{
\code{\link{plot_umap}} \code{\link{plot_umap_heat}} \code{\link{plot_density_umap}}
}
