% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{select_files_shiny}
\alias{select_files_shiny}
\title{Select files for flow object - Shiny}
\usage{
select_files_shiny(dir)
}
\arguments{
\item{dir}{working directory}
}
\value{
A file list
}
\description{
Select files
}
