% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SubsetFlowObject}
\alias{SubsetFlowObject}
\title{Subset a FlowObject.}
\usage{
SubsetFlowObject(flow_object, subset)
}
\arguments{
\item{flow_object}{A flow_object}

\item{subset}{criteria for filtering. See Details.}
}
\value{
A flow_object
}
\description{
Subset a FlowObject on metadata criteria, cluster annotations, dimensions or marker fluorescence.
}
\details{
Any parameter embedded in the flow_object can be potentially used as a filter, as a logical expression.
This includes marker fluorescence expression, cluster name or annotation, sample metadata values, UMAP coordinates.
Logical expressions can be chained together using the & (AND) or | operators (OR) .

Expression operators such as == (equals to), != (does not equal), > or < (greater or smaller than), %in% (is a part of ...) are valid.
}
