% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_trajectory-data.R
\docType{data}
\name{th_trajectory}
\alias{th_trajectory}
\title{CD4 Th differentiation flow object}
\format{
An object of class \code{flow_object} of length 7.
}
\usage{
data(th_trajectory)
}
\description{
Data from a pregated CD4 T cell staining `(CD3+CD4+)`. Cells were TCR stimulated, and different conditions were given distinct Th differentiating cocktails.
}
\keyword{datasets}
