% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{UMAP_flow}
\alias{UMAP_flow}
\title{UMAP Dimension Reduction}
\usage{
UMAP_flow(
  flow_object,
  transformation = "asinh",
  n_neighbors = 15,
  scale = TRUE,
  dims = 2
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{transformation}{Data transformation method to apply to the data. Possible choices are 'asinh', 'none', 'log' or 'cytofAsinh'. See `details` for more information}

\item{n_neighbors}{Number of neighbors. See documentation from the `uwot::umap()` function for more details.}

\item{scale}{Logical argument to determine whether or not to scale the data prior to umap. Defaults to TRUE}

\item{dims}{Number of dimensions to reduce the data unto. 2 is the best for the vast majority of cases, but 3 can be useful for trajectory analysis.}
}
\value{
A Flow Object
}
\description{
Runs the Uniform Manifold Approximation and Projection (UMAP) dimension reduction technique on the flow cytometry data.
}
\details{
Possible choices for data transformation are hyperbolic sine (asinh)
using cofactors inferred by `find_cofactors()`, biexponential (none), logarithmic (in base 10), or CYTOF asinh (cofactor of 5).
}
