% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_sample_metadata}
\alias{add_sample_metadata}
\title{Add metadata annotations}
\usage{
add_sample_metadata(flow_object, metadata, key_column)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{metadata}{A data.frame containing sample annotations. Must contain a column called `SampleID`, that contains the sample names stored in the Flow Object.}

\item{key_column}{Name of the column matching sample names.}
}
\value{
A Flow Object
}
\description{
Add sample annotations to the Flow Object. This is where you add clinical variables (i.e. antibody levels, Timepoints, Treatment, etc) or technical variables (like Batch).
Variables can be discrete (i.e. categorical) or continuous. If using categorical, it is recommended to assign factor levels if you desire a specific order for labels to be displayed in plots, or how statistics will be performed.
By default, the order will be performed in alphabetical order.
}
