% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{flow_to_cds}
\alias{flow_to_cds}
\title{Export Flow Object to a monocle 3 CDS object.}
\usage{
flow_to_cds(
  flow_object,
  partition,
  downsample = 5000,
  min = 200,
  scale = FALSE
)
}
\arguments{
\item{flow_object}{A Flow Object}

\item{partition}{Name of a partition in the Flow Object to perform the trajectory analysis with. Partitions are first set to '1' or using the 'SetPartitions' function.}

\item{downsample}{Number of events to keep PER CLUSTER. Cluster based downsampling helps to reducing the computational weight of high frequency clusters, especially if using high subsampling numbers at object creation.}

\item{min}{Minimum number of events to take per cluster.}

\item{scale}{Logical argument to determine whether or not to scale the data prior to automatic detection of partitions.}
}
\value{
A Flow Object with partition annotations.
}
\description{
Monocle 3 attempts to detect separate groups of cells that likely fall within the same trajectories. By default, all cells are set up to be part of the same
partition (named "1"), but this behavior does not always make biological sense. For instance, when analyzing whole blood PBMCs, a trajectory between monocytes and T or B cells is not expected.
This will detect them as 2 separate partitions and will make them to be analyzable in parallel.
Partitions can also be manually set by providing a data.frame of partitions for each cell, with a 'cellID' column.
}
